[0.0.10] Improved Texture Loading 🌠

+++ core/config.h
@@ -23,7 +23,7 @@ namespace Verse
         int palette_index;
         int num_palettes;
         
-        int background_color[4];
+        float background_color[4];
         
         bool render_collision_boxes;
     };
